requirements.txt
absl-py==2.1.0
altair==5.4.1
astunparse==1.6.3
attrs==24.2.0
blinker==1.8.1
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
filelock==3.16.1
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.68.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.3
joblib==1.4.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.6.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
namex==0.0.8
narwhals==1.13.5
networkx==3.4.1
numpy==1.26.4
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
protobuf==5.28.2
psutil==6.0.0
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyaml==24.9.0
pyarrow==18.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.13.0
seaborn==0.13.2
setuptools==69.5.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.29
statsmodels==0.14.2
streamlit==1.40.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
toml==0.10.2
torch==2.5.0
torchvision==0.20.0
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.11.0
tzdata==2024.1
ultralytics-thop==2.0.9
urllib3==2.2.3
Werkzeug==3.0.2
wheel==0.43.0
wrapt==1.17.0
